#75: Rename Enum.h to UsableEnum.h
