+++ bask3twork/MainWindow.cpp
@@ -3,10 +3,10 @@
 #include "MainWindow.h"
 #include "grid/Display.h"
 #include "grid/Knot.h"
-#include "pure/Enum.h"
 #include "pure/Glyph.h"
 #include "pure/GridSize.h"
 #include "pure/Symmetry.h"
+#include "pure/UsableEnum.h"
 #include "regions/Select.h"
 #include "regions/Generate.h"
 #include "regions/Export.h"
