+++ bask3twork/bask3twork.vcxproj
@@ -42,7 +42,7 @@
     <ClInclude Include="pch.h" />
     <ClInclude Include="pure\Connection.h" />
     <ClInclude Include="grid\Display.h" />
-    <ClInclude Include="pure\Enum.h" />
+    <ClInclude Include="pure\UsableEnum.h" />
     <ClInclude Include="pure\GridSize.h" />
     <ClInclude Include="regions\Export.h" />
     <ClInclude Include="Forward.h" />
