+++ bask3twork/bask3twork.vcxproj.filters
@@ -101,7 +101,7 @@
     <ClInclude Include="controls\RegenDialog.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="pure\Enum.h">
+    <ClInclude Include="pure\UsableEnum.h">
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="Forward.h">
