+++ bask3twork/pure/UsableEnum.h
